% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{createConnectionParameters}
\alias{createConnectionParameters}
\title{createConnectionParameters}
\usage{
createConnectionParameters(dbms, host, dBname = NA, dBsourceSchema,
  dBtargetSchema, user, password)
}
\arguments{
\item{dbms}{one of "postgresql", "sqlserver", "oracle *", "mysql *", or "redshift *" (* not yet implemented)}

\item{host}{the ip address/host name of the database instance}

\item{dBname}{for PostgreSQL database only: the database name}

\item{dBsourceSchema}{the source database schema}

\item{dBtargetSchema}{the target database schema}

\item{user}{the user name}

\item{password}{the user password}
}
\value{
A string JSON object containing details for connecting to the database
}
\description{
createConnectionParameters
}
\details{
Create connection parameters json from a set of parameters.
}
