% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{exportCohortResults}
\alias{exportCohortResults}
\title{exportCohortResults}
\usage{
exportCohortResults(connectionParameters, outFilePath)
}
\arguments{
\item{connectionParameters}{a string JSON object containing details for connecting to the database}

\item{outFilePath}{file path to where the results should be written}
}
\description{
exportCohortResults
}
\details{
Extracts the cohort_inclusion, cohort_inclusion_result, cohort_inclusion_stats,
and cohort_summary_stats records for a cohort definition.
Saves the resulting cohort result records as a JSON file.
}
