% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{exportS3CohortResults}
\alias{exportS3CohortResults}
\title{exportS3CohortResults}
\usage{
exportS3CohortResults(connectionParameters, bucketName, s3keys)
}
\arguments{
\item{connectionParameters}{a string JSON object containing details for connecting to the database.}

\item{bucketName}{name of the bucket on S3 where to save the cohort inclusion report result file.}

\item{s3keys}{file path to the csv file with the S3 access keys}
}
\description{
exportS3CohortResults
}
\details{
Extracts the cohort_inclusion, cohort_inclusion_result, cohort_inclusion_stats,
and cohort_summary_stats records for a cohort definition.
Saves the resulting cohort result records as a JSON file in the specified S3 bucket.
Export file is named 'cohort_incl_res_xx.json', where xx is the local cohort ID.
}
