% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{getCohortResults}
\alias{getCohortResults}
\title{getCohortResults}
\usage{
getCohortResults(connectionParameters, cohortID)
}
\arguments{
\item{connectionParameters}{a string JSON object containing details for connecting to the database}

\item{cohortID}{the cohort_definition_id for which to extract the results}
}
\value{
JSON string with results from the executed SQL script
}
\description{
getCohortResults
}
\details{
Returns the cohort result records for the specified cohort in a JSON structure.
}
