% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{getConnectionParameters}
\alias{getConnectionParameters}
\title{getConnectionParameters}
\usage{
getConnectionParameters(connectionFile)
}
\arguments{
\item{connectionFile}{the full path to the json file}
}
\value{
A string JSON object containing details for connecting to the database
}
\description{
getConnectionParameters
}
\details{
Loads connection parameters from a json file.

The JSON file should have the following format:

\code{[ \cr
   {	\cr
     "dbms":"<dbms>",\cr
     "server":"<host-ip>",\cr
     "dbname":"<PostgreSQL db name>",\cr
     "user":"<user>",\cr
     "password":"<password>",\cr
     "sourceschema":"<database schema>",\cr
     "targetschema":"<database schema>"\cr
  }\cr
] \cr
}

\code{dbms} can be one of "postgresql", "sqlserver", "oracle *", "mysql *", or "redshift *" (* not yet implemented)\cr
\code{dbname} is only applicable for PostgreSQL databases
}
