% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{importCohortResults}
\alias{importCohortResults}
\title{importCohortResults}
\usage{
importCohortResults(connectionParameters, inFilePath)
}
\arguments{
\item{connectionParameters}{a string JSON object containing details for connecting to the database}

\item{inFilePath}{file path to the cohort inclusion results JSON file}
}
\description{
importCohortResults
}
\details{
Loads the cohort inclusion results JSON file, and uploads/imports
the records to the corresponding tables in the specified schema
with the specified cohort ID.
NOTE: INCOMPLETE!! Currently only works when the cohort results tables are in the default schema!
}
