% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{insertCohortDefinition}
\alias{insertCohortDefinition}
\title{insertCohortDefinition}
\usage{
insertCohortDefinition(connectionParameters, jsonPath = NA)
}
\arguments{
\item{connectionParameters}{a string JSON object containing details for connecting to the database}

\item{jsonPath}{the path to a JSON file containing the cohort definition - if empty, prompts for file}
}
\description{
insertCohortDefinition
}
\details{
Inserts the cohort definition records from the provided JSON file into the relevant tables.
}
