% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasCohortExchange.R
\name{readinteger}
\alias{readinteger}
\title{readinteger(promptStr)}
\usage{
readinteger(promptStr)
}
\arguments{
\item{promptStr}{The string to display when asking user for input.}
}
\value{
An integer.
}
\description{
readinteger(promptStr)
}
\details{
Prompts the user to enter an integer
}
